Extracting Add and Preview Transform logic
