+++ src/Microsoft.VisualStudio.SlowCheetah.VS/SlowCheetahPackage.cs
@@ -86,25 +86,6 @@ namespace Microsoft.VisualStudio.SlowCheetah.VS
         private SlowCheetahNuGetManager NuGetManager { get; }
 
         /// <summary>
-        /// Gets the installation directory for the current instance of Visual Studio.
-        /// </summary>
-        /// <returns>Full path to the VS instalation directory</returns>
-        public string GetVsInstallDirectory()
-        {
-            string installDirectory = null;
-            if (this.GetService(typeof(SVsShell)) is IVsShell shell)
-            {
-                shell.GetProperty((int)__VSSPROPID.VSSPROPID_InstallDirectory, out object installDirectoryObj);
-                if (installDirectoryObj != null)
-                {
-                    installDirectory = installDirectoryObj as string;
-                }
-            }
-
-            return installDirectory;
-        }
-
-        /// <summary>
         /// Initialization of the package; this method is called right after the package is sited, so this is the place
         /// where you can put all the initialization code that rely on services provided by VisualStudio.
         /// </summary>
@@ -160,322 +141,6 @@ namespace Microsoft.VisualStudio.SlowCheetah.VS
             base.Dispose(disposing);
         }
 
-        private void OnChangeAddTransformMenu(object sender, EventArgs e)
-        {
-        }
-
-        private void OnChangePreviewTransformMenu(object sender, EventArgs e)
-        {
-        }
-
-        /// <summary>
-        /// This event is fired when a user right-clicks on a menu, but prior to the menu showing. This function is used to set the visibility
-        /// of the "Add Transform" menu. It checks to see if the project is one of the supported types, and if the extension of the project item
-        /// that was right-clicked on is one of the valid item types.
-        /// </summary>
-        /// <param name="sender">The menu that fired the event.</param>
-        /// <param name="e">Not used.</param>
-        private void OnBeforeQueryStatusAddTransformCommand(object sender, EventArgs e)
-        {
-            // get the menu that fired the event
-            if (sender is OleMenuCommand menuCommand)
-            {
-                // start by assuming that the menu will not be shown
-                menuCommand.Visible = false;
-                menuCommand.Enabled = false;
-                uint itemid = VSConstants.VSITEMID_NIL;
-
-                if (!ProjectUtilities.IsSingleProjectItemSelection(out IVsHierarchy hierarchy, out itemid))
-                {
-                    return;
-                }
-
-                IVsProject vsProject = (IVsProject)hierarchy;
-                if (!this.ProjectSupportsTransforms(vsProject))
-                {
-                    return;
-                }
-
-                if (!this.ItemSupportsTransforms(vsProject, itemid))
-                {
-                    return;
-                }
-
-                menuCommand.Visible = true;
-                menuCommand.Enabled = true;
-            }
-        }
-
-        /// <summary>
-        /// This event is fired when a user right-clicks on a menu, but prior to the menu showing. This function is used to set the visibility
-        /// of the "Add Transform" menu. It checks to see if the project is one of the supported types, and if the extension of the project item
-        /// that was right-clicked on is one of the valid item types.
-        /// </summary>
-        /// <param name="sender">The menu that fired the event.</param>
-        /// <param name="e">Not used.</param>
-        private void OnBeforeQueryStatusPreviewTransformCommand(object sender, EventArgs e)
-        {
-            // get the menu that fired the event
-            if (sender is OleMenuCommand menuCommand)
-            {
-                // start by assuming that the menu will not be shown
-                menuCommand.Visible = false;
-                menuCommand.Enabled = false;
-                uint itemid = VSConstants.VSITEMID_NIL;
-
-                if (!ProjectUtilities.IsSingleProjectItemSelection(out IVsHierarchy hierarchy, out itemid))
-                {
-                    return;
-                }
-
-                IVsProject vsProject = (IVsProject)hierarchy;
-                if (!this.ProjectSupportsTransforms(vsProject))
-                {
-                    return;
-                }
-
-                if (!this.IsItemTransformItem(vsProject, itemid))
-                {
-                    return;
-                }
-
-                menuCommand.Visible = true;
-                menuCommand.Enabled = true;
-            }
-        }
-
-        /// <summary>
-        /// This function is the callback used to execute a command when the a menu item is clicked.
-        /// See the Initialize method to see how the menu item is associated to this function using
-        /// the OleMenuCommandService service and the MenuCommand class.
-        /// </summary>
-        /// <param name="sender">The object that fired the event</param>
-        /// <param name="e">Event arguments</param>
-        private void OnAddTransformCommand(object sender, EventArgs e)
-        {
-            uint itemid = VSConstants.VSITEMID_NIL;
-
-            if (!ProjectUtilities.IsSingleProjectItemSelection(out IVsHierarchy hierarchy, out itemid))
-            {
-                return;
-            }
-
-            IVsProject vsProject = (IVsProject)hierarchy;
-            if (!this.ProjectSupportsTransforms(vsProject))
-            {
-                return;
-            }
-
-            if (ErrorHandler.Failed(vsProject.GetMkDocument(VSConstants.VSITEMID_ROOT, out string projectFullPath)))
-            {
-                return;
-            }
-
-            IVsBuildPropertyStorage buildPropertyStorage = vsProject as IVsBuildPropertyStorage;
-            if (buildPropertyStorage == null)
-            {
-                this.LogMessageWriteLineFormat("Error obtaining IVsBuildPropertyStorage from hierarchy.");
-                return;
-            }
-
-            // get the name of the item
-            if (ErrorHandler.Failed(vsProject.GetMkDocument(itemid, out string itemFullPath)))
-            {
-                return;
-            }
-
-            // Save the project file
-            IVsSolution solution = (IVsSolution)Package.GetGlobalService(typeof(SVsSolution));
-            int hr = solution.SaveSolutionElement((uint)__VSSLNSAVEOPTIONS.SLNSAVEOPT_SaveIfDirty, hierarchy, 0);
-            if (ErrorHandler.Failed(hr))
-            {
-                throw new COMException(string.Format(Resources.Resources.Error_SavingProjectFile, itemFullPath, this.GetErrorInfo()), hr);
-            }
-
-            ProjectItem selectedProjectItem = PackageUtilities.GetAutomationFromHierarchy<ProjectItem>(hierarchy, itemid);
-            if (selectedProjectItem != null)
-            {
-                try
-                {
-                    selectedProjectItem.Save();
-                }
-                catch
-                {
-                    // If the item is not open, an exception is thrown,
-                    // but that is not a problem as it is not dirty
-                }
-
-                // Checks the SlowCheetah NuGet package installation
-                this.NuGetManager.CheckSlowCheetahInstallation(hierarchy);
-
-                // need to enure that this item has metadata TransformOnBuild set to true
-                buildPropertyStorage.SetItemAttribute(itemid, TransformOnBuild, "true");
-
-                string itemFolder = Path.GetDirectoryName(itemFullPath);
-                string itemFilename = Path.GetFileNameWithoutExtension(itemFullPath);
-                string itemExtension = Path.GetExtension(itemFullPath);
-                string itemFilenameExtension = Path.GetFileName(itemFullPath);
-
-                IEnumerable<string> configs = ProjectUtilities.GetProjectConfigurations(selectedProjectItem.ContainingProject);
-
-                List<string> transformsToCreate = null;
-                if (configs != null)
-                {
-                    transformsToCreate = configs.ToList();
-                }
-
-                if (transformsToCreate == null)
-                {
-                    transformsToCreate = new List<string>();
-                }
-
-                // if it is a web project we should add publish profile specific transforms as well
-                var publishProfileTransforms = this.GetPublishProfileTransforms(hierarchy, projectFullPath);
-                if (publishProfileTransforms != null)
-                {
-                    transformsToCreate.AddRange(publishProfileTransforms);
-                }
-
-                using (OptionsDialogPage optionsPage = new OptionsDialogPage())
-                {
-                    optionsPage.LoadSettingsFromStorage();
-
-                    foreach (string config in transformsToCreate)
-                    {
-                        string itemName = string.Format(Resources.Resources.String_FormatTransformFilename, itemFilename, config, itemExtension);
-                        this.AddTransformFile(selectedProjectItem, itemName, itemFolder, optionsPage.AddDependentUpon);
-                        hierarchy.ParseCanonicalName(Path.Combine(itemFolder, itemName), out uint addedFileId);
-                        buildPropertyStorage.SetItemAttribute(addedFileId, IsTransformFile, "True");
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// This function is the callback used to execute a command when the a menu item is clicked.
-        /// See the Initialize method to see how the menu item is associated to this function using
-        /// the OleMenuCommandService service and the MenuCommand class.
-        /// </summary>
-        /// <param name="sender">>The object that fired the event</param>
-        /// <param name="e">Event arguments</param>
-        private void OnPreviewTransformCommand(object sender, EventArgs e)
-        {
-            uint itemId = VSConstants.VSITEMID_NIL;
-
-            // verify only one item is selected
-            if (!ProjectUtilities.IsSingleProjectItemSelection(out IVsHierarchy hierarchy, out itemId))
-            {
-                return;
-            }
-
-            // make sure that the SlowCheetah project support has been added
-            IVsProject project = (IVsProject)hierarchy;
-            if (!this.ProjectSupportsTransforms(project))
-            {
-                // TODO: should add a dialog here telling the user that the preview failed because the targets are not yet installed
-                return;
-            }
-
-            // get the full path of the configuration xdt
-            if (ErrorHandler.Failed(project.GetMkDocument(itemId, out string transformPath)))
-            {
-                return;
-            }
-
-            // Checks the SlowCheetah NuGet package installation
-            this.NuGetManager.CheckSlowCheetahInstallation(hierarchy);
-
-            ErrorHandler.ThrowOnFailure(hierarchy.GetProperty(itemId, (int)__VSHPROPID.VSHPROPID_Parent, out object parentIdObj));
-            uint parentId = (uint)(int)parentIdObj;
-            if (parentId == (uint)VSConstants.VSITEMID.Nil)
-            {
-                return;
-            }
-
-            if (!this.TryGetFileToTransform(hierarchy, parentId, Path.GetFileName(transformPath), out uint docId, out string documentPath))
-            {
-                // TO DO: Possibly tell the user that the transform file was not found.
-                return;
-            }
-
-            try
-            {
-                PackageUtilities.GetAutomationFromHierarchy<ProjectItem>(hierarchy, docId).Save();
-                PackageUtilities.GetAutomationFromHierarchy<ProjectItem>(hierarchy, itemId).Save();
-            }
-            catch
-            {
-                // If the item is not open, an exception is thrown,
-                // but that is not a problem as it is not dirty
-            }
-
-            this.PreviewTransform(hierarchy, documentPath, transformPath);
-        }
-
-        /// <summary>
-        /// Searches for a file to transform based on a transformation file.
-        /// Starts the search with the parent of the file then checks all visible children.
-        /// </summary>
-        /// <param name="hierarchy">Current project hierarchy</param>
-        /// <param name="parentId">Parent ID of the file.</param>
-        /// <param name="transformName">Name of the transformation file</param>
-        /// <param name="docId">ID of the file to transform</param>
-        /// <param name="documentPath">Resulting path of the file to transform</param>
-        /// <returns>True if the correct file was found</returns>
-        private bool TryGetFileToTransform(IVsHierarchy hierarchy, uint parentId, string transformName, out uint docId, out string documentPath)
-        {
-            IVsProject project = (IVsProject)hierarchy;
-
-            IEnumerable<string> configs = ProjectUtilities.GetProjectConfigurations(hierarchy);
-
-            if (ErrorHandler.Failed(project.GetMkDocument(parentId, out documentPath)))
-            {
-                docId = 0;
-                return false;
-            }
-
-            if (PackageUtilities.IsFileTransform(Path.GetFileName(documentPath), transformName, configs))
-            {
-                docId = parentId;
-                return true;
-            }
-            else
-            {
-                hierarchy.GetProperty(parentId, (int)__VSHPROPID.VSHPROPID_FirstVisibleChild, out object childIdObj);
-                docId = (uint)(int)childIdObj;
-                if (ErrorHandler.Failed(project.GetMkDocument(docId, out documentPath)))
-                {
-                    docId = 0;
-                    documentPath = null;
-                    return false;
-                }
-
-                if (PackageUtilities.IsFileTransform(Path.GetFileName(documentPath), transformName, configs))
-                {
-                    return true;
-                }
-                else
-                {
-                    while (docId != VSConstants.VSITEMID_NIL)
-                    {
-                        hierarchy.GetProperty(docId, (int)__VSHPROPID.VSHPROPID_NextVisibleSibling, out childIdObj);
-                        docId = (uint)(int)childIdObj;
-                        if (ErrorHandler.Succeeded(project.GetMkDocument(docId, out documentPath)))
-                        {
-                            if (PackageUtilities.IsFileTransform(Path.GetFileName(documentPath), transformName, configs))
-                            {
-                                return true;
-                            }
-                        }
-                    }
-                }
-            }
-
-            docId = 0;
-            documentPath = null;
-            return false;
-        }
-
         /// <summary>
         /// Verifies if the item has a trasform configured already
         /// </summary>
@@ -506,48 +171,6 @@ namespace Microsoft.VisualStudio.SlowCheetah.VS
         }
 
         /// <summary>
-        /// Verifies any publish profiles in the project and returns it as a list of strings
-        /// </summary>
-        /// <param name="hierarchy">The current project hierarchy</param>
-        /// <param name="projectPath">Full path of the current project</param>
-        /// <returns>List of publish profile names</returns>
-        private IEnumerable<string> GetPublishProfileTransforms(IVsHierarchy hierarchy, string projectPath)
-        {
-            if (hierarchy == null)
-            {
-                throw new ArgumentNullException(nameof(hierarchy));
-            }
-
-            if (string.IsNullOrEmpty(projectPath))
-            {
-                throw new ArgumentNullException(nameof(projectPath));
-            }
-
-            List<string> result = new List<string>();
-            IVsProjectSpecialFiles specialFiles = hierarchy as IVsProjectSpecialFiles;
-            if (ErrorHandler.Failed(specialFiles.GetFile((int)__PSFFILEID2.PSFFILEID_AppDesigner, (uint)__PSFFLAGS.PSFF_FullPath, out uint itemid, out string propertiesFolder)))
-            {
-                this.LogMessageWriteLineFormat("Exception trying to create IVsProjectSpecialFiles");
-            }
-
-            if (!string.IsNullOrEmpty(propertiesFolder))
-            {
-                // Properties\PublishProfiles
-                string publishProfilesFolder = Path.Combine(propertiesFolder, "PublishProfiles");
-                if (Directory.Exists(publishProfilesFolder))
-                {
-                    string[] publishProfiles = Directory.GetFiles(publishProfilesFolder, "*.pubxml");
-                    if (publishProfiles != null)
-                    {
-                        return publishProfiles.Select(profile => Path.GetFileNameWithoutExtension(profile));
-                    }
-                }
-            }
-
-            return null;
-        }
-
-        /// <summary>
         /// Verifies if the current project supports transformations.
         /// </summary>
         /// <param name="project">Current IVsProject</param>
@@ -555,177 +178,6 @@ namespace Microsoft.VisualStudio.SlowCheetah.VS
         private bool ProjectSupportsTransforms(IVsProject project)
         {
             return this.NuGetManager.ProjectSupportsNuget(project as IVsHierarchy);
-        }
-
-        /// <summary>
-        /// Verifies if an item supports transforms.
-        /// </summary>
-        /// <param name="project">Current IVsProject</param>
-        /// <param name="itemid">Id of the item inside the project</param>
-        /// <returns>True if the item supports transforms</returns>
-        private bool ItemSupportsTransforms(IVsProject project, uint itemid)
-        {
-            if (ErrorHandler.Failed(project.GetMkDocument(itemid, out string itemFullPath)))
-            {
-                return false;
-            }
-
-            bool itemSupportsTransforms = false;
-            FileInfo transformFileInfo = new FileInfo(itemFullPath);
-
-            // make sure its not a transform file itself
-            bool isWebConfig = string.Compare("web.config", transformFileInfo.Name, StringComparison.OrdinalIgnoreCase) == 0;
-            bool isTransformFile = this.IsItemTransformItem(project, itemid);
-            bool isExtensionSupportedForFile = PackageUtilities.IsExtensionSupportedForFile(itemFullPath);
-            bool isSupportedFile = TransformerFactory.IsSupportedFile(itemFullPath);
-
-            if (!isWebConfig && !isTransformFile && isExtensionSupportedForFile && isSupportedFile)
-            {
-                itemSupportsTransforms = true;
-            }
-
-            return itemSupportsTransforms;
-        }
-
-        /// <summary>
-        /// Creates a new transformation file and adds it to the project.
-        /// </summary>
-        /// <param name="selectedProjectItem">The selected item to be transformed</param>
-        /// <param name="itemName">Full name of the transformation file</param>
-        /// <param name="projectPath">Full path to the current project</param>
-        /// <param name="addDependentUpon">Wheter to add the new file dependent upon the source file</param>
-        private void AddTransformFile(ProjectItem selectedProjectItem, string itemName, string projectPath, bool addDependentUpon)
-        {
-            try
-            {
-                string transformPath = Path.Combine(projectPath, itemName);
-                string sourceFileName = selectedProjectItem.FileNames[1];
-
-                ITransformer transformer = TransformerFactory.GetTransformer(sourceFileName, null);
-
-                transformer.CreateTransformFile(sourceFileName, transformPath, false);
-
-                // Add the file to the project
-                // If the DependentUpon metadata is required, add it under the original file
-                // If not, add it to the project
-                ProjectItem addedItem = addDependentUpon ? selectedProjectItem.ProjectItems.AddFromFile(transformPath)
-                                                      : selectedProjectItem.ContainingProject.ProjectItems.AddFromFile(transformPath);
-
-                // we need to set the Build Action to None to ensure that it doesn't get published for web projects
-                addedItem.Properties.Item("ItemType").Value = "None";
-
-                IVsHierarchy hierarchy = null;
-                IVsProject vsProject = (IVsProject)hierarchy;
-                IVsBuildPropertyStorage buildPropertyStorage = vsProject as IVsBuildPropertyStorage;
-                if (buildPropertyStorage == null)
-                {
-                    this.LogMessageWriteLineFormat("Error obtaining IVsBuildPropertyStorage from hierarcy.");
-                }
-            }
-            catch (Exception ex)
-            {
-                this.LogMessageWriteLineFormat("AddTransformFile: Exception> " + ex.Message);
-            }
-        }
-
-        /// <summary>
-        /// Shows a preview of the transformation in a temporary file.
-        /// </summary>
-        /// <param name="hier">Current IVsHierarchy</param>
-        /// <param name="sourceFile">Full path to the file to be transformed</param>
-        /// <param name="transformFile">Full path to the transformation file</param>
-        private void PreviewTransform(IVsHierarchy hier, string sourceFile, string transformFile)
-        {
-            if (string.IsNullOrWhiteSpace(sourceFile))
-            {
-                throw new ArgumentNullException(nameof(sourceFile));
-            }
-
-            if (string.IsNullOrWhiteSpace(transformFile))
-            {
-                throw new ArgumentNullException(nameof(transformFile));
-            }
-
-            if (!File.Exists(sourceFile))
-            {
-                throw new FileNotFoundException(string.Format(CultureInfo.CurrentCulture, Resources.Resources.Error_SourceFileNotFound, sourceFile), sourceFile);
-            }
-
-            if (!File.Exists(transformFile))
-            {
-                throw new FileNotFoundException(string.Format(CultureInfo.CurrentCulture, Resources.Resources.Error_TransformFileNotFound, transformFile), transformFile);
-            }
-
-            // Get our options
-            using (OptionsDialogPage optionsPage = new OptionsDialogPage())
-            using (AdvancedOptionsDialogPage advancedOptionsPage = new AdvancedOptionsDialogPage())
-            {
-                optionsPage.LoadSettingsFromStorage();
-                advancedOptionsPage.LoadSettingsFromStorage();
-
-                this.LogMessageWriteLineFormat("SlowCheetah PreviewTransform");
-                FileInfo sourceFileInfo = new FileInfo(sourceFile);
-
-                // dest file
-                string destFile = PackageUtilities.GetTempFilename(true, sourceFileInfo.Extension);
-                this.TempFilesCreated.Add(destFile);
-
-                // perform the transform and then display the result into the diffmerge tool that comes with VS.
-                // If for some reason we can't find it, we just open it in an editor window
-                this.errorListProvider.Tasks.Clear();
-                ITransformationLogger logger = new TransformationPreviewLogger(this.errorListProvider, hier);
-                ITransformer transformer = TransformerFactory.GetTransformer(sourceFile, logger);
-                if (!transformer.Transform(sourceFile, transformFile, destFile))
-                {
-                    throw new TransformFailedException(Resources.Resources.TransformPreview_ErrorMessage);
-                }
-
-                // Does the customer want a preview?
-                if (optionsPage.EnablePreview == false)
-                {
-                    ProjectUtilities.GetDTE().ItemOperations.OpenFile(destFile);
-                }
-                else
-                {
-                    // If the diffmerge service is available (dev11) and no diff tool is specified, or diffmerge.exe is specifed we use the service
-                    if (this.GetService(typeof(SVsDifferenceService)) is IVsDifferenceService diffService && (!File.Exists(advancedOptionsPage.PreviewToolExecutablePath) || advancedOptionsPage.PreviewToolExecutablePath.EndsWith("diffmerge.exe", StringComparison.OrdinalIgnoreCase)))
-                    {
-                        if (!string.IsNullOrEmpty(advancedOptionsPage.PreviewToolExecutablePath) && !File.Exists(advancedOptionsPage.PreviewToolExecutablePath))
-                        {
-                            logger.LogWarning(string.Format(Resources.Resources.Error_CantFindPreviewTool, advancedOptionsPage.PreviewToolExecutablePath));
-                        }
-
-                        string sourceName = Path.GetFileName(sourceFile);
-                        string leftLabel = string.Format(CultureInfo.CurrentCulture, Resources.Resources.TransformPreview_LeftLabel, sourceName);
-                        string rightLabel = string.Format(CultureInfo.CurrentCulture, Resources.Resources.TransformPreview_RightLabel, sourceName, Path.GetFileName(transformFile));
-                        string caption = string.Format(CultureInfo.CurrentCulture, Resources.Resources.TransformPreview_Caption, sourceName);
-                        string tooltip = string.Format(CultureInfo.CurrentCulture, Resources.Resources.TransformPreview_ToolTip, sourceName);
-                        diffService.OpenComparisonWindow2(sourceFile, destFile, caption, tooltip, leftLabel, rightLabel, null, null, (uint)__VSDIFFSERVICEOPTIONS.VSDIFFOPT_RightFileIsTemporary);
-                    }
-                    else if (string.IsNullOrEmpty(advancedOptionsPage.PreviewToolExecutablePath))
-                    {
-                        throw new FileNotFoundException(Resources.Resources.Error_NoPreviewToolSpecified);
-                    }
-                    else if (!File.Exists(advancedOptionsPage.PreviewToolExecutablePath))
-                    {
-                        throw new FileNotFoundException(string.Format(Resources.Resources.Error_CantFindPreviewTool, advancedOptionsPage.PreviewToolExecutablePath), advancedOptionsPage.PreviewToolExecutablePath);
-                    }
-                    else
-                    {
-                        // Quote the filenames...
-                        ProcessStartInfo psi = new ProcessStartInfo(advancedOptionsPage.PreviewToolExecutablePath, string.Format(advancedOptionsPage.PreviewToolCommandLine, $"\"{sourceFile}\"", $"\"{destFile}\""))
-                        {
-                            CreateNoWindow = true,
-                            UseShellExecute = false
-                        };
-                        System.Diagnostics.Process.Start(psi);
-                    }
-                }
-            }
-
-            // TODO: Instead of creating a file and then deleting it later we could instead do this
-            //          http://matthewmanela.com/blog/the-problem-with-the-envdte-itemoperations-newfile-method/
-            //          http://social.msdn.microsoft.com/Forums/en/vsx/thread/eb032063-eb4d-42e0-84e8-dec64bf42abf
         }
 
         /// <summary>
